изменил название "label fragment" nav_graph ( отображение названия в ActionBar)
