+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.example.redditpopularapi"
+        applicationId "com.android.lookerpopularapi"
         minSdk 21
         targetSdk 32
         versionCode 1
