+++ app/src/main/java/com/android/lookerpopularapi/data/models/childrenModels/Children.kt
@@ -1,0 +1,10 @@
+package com.android.lookerpopularapi.data.models.childrenModels
+
+
+import com.google.gson.annotations.SerializedName
+
+data class Children(
+    @SerializedName("data")
+    val childrenData: ChildrenData,
+
+    )
