+++ app/src/main/java/com/android/lookerpopularapi/data/models/childrenModels/ChildrenData.kt
@@ -1,0 +1,23 @@
+package com.android.lookerpopularapi.data.models.childrenModels
+
+
+import com.google.gson.annotations.SerializedName
+
+data class ChildrenData(
+
+    @SerializedName("author_fullname")
+    val authorFullname: String,
+    @SerializedName("title")
+    val title: String,
+    @SerializedName("thumbnail")
+    val thumbnail: String,
+    @SerializedName("created")
+    val created: Double,
+    @SerializedName("num_comments")
+    val numComments: Int,
+    @SerializedName("total_awards_received")
+    val totalAwardsReceived: Int,
+    @SerializedName("name")
+    val name: String,
+
+)
