+++ app/src/main/java/com/android/lookerpopularapi/data/models/popularItemModels/PopularItemRedditModel.kt
@@ -1,0 +1,11 @@
+package com.android.lookerpopularapi.data.models.popularItemModels
+
+
+import com.google.gson.annotations.SerializedName
+
+data class PopularItemRedditModel(
+    @SerializedName("data")
+    val `data`: Data,
+    @SerializedName("kind")
+    val kind: String
+)
