+++ app/src/main/java/com/android/lookerpopularapi/data/source/AppState.kt
@@ -1,0 +1,7 @@
+package com.android.lookerpopularapi.data.source
+
+sealed class AppState{
+    data class Success(val data: Any) : AppState()
+    data class Error(val error_ : Throwable) : AppState()
+    object Loading: AppState()
+}
