+++ app/src/main/java/com/android/lookerpopularapi/data/source/remote/ApiSource.kt
@@ -1,0 +1,8 @@
+package com.android.lookerpopularapi.data.source.remote
+
+import com.android.lookerpopularapi.data.models.popularItemModels.PopularItemRedditModel
+
+interface ApiSource {
+    suspend fun getData(): PopularItemRedditModel
+    suspend fun getMoreData(after: String): PopularItemRedditModel
+}
