+++ app/src/main/java/com/android/lookerpopularapi/data/source/remote/api/OkhttpClientInstance.kt
@@ -1,0 +1,17 @@
+package com.android.lookerpopularapi.data.source.remote.api
+
+import okhttp3.OkHttpClient
+
+object OkhttpClientInstance {
+    fun getOkHttpClientBuilder(): OkHttpClient {
+        val httpClient = OkHttpClient.Builder()
+            .addInterceptor { chain ->
+                val original = chain.request()
+                val request = original.newBuilder()
+                    .method(original.method, original.body)
+                    .build()
+                chain.proceed(request)
+            }
+        return httpClient.build()
+    }
+}
