+++ app/src/main/java/com/android/lookerpopularapi/data/source/repository/Repository.kt
@@ -1,0 +1,12 @@
+package com.android.lookerpopularapi.data.source.repository
+
+import androidx.paging.PagingData
+import com.android.lookerpopularapi.data.models.childrenModels.Children
+import com.android.lookerpopularapi.data.models.popularItemModels.PopularItemRedditModel
+
+import kotlinx.coroutines.flow.Flow
+
+interface Repository {
+    suspend fun getData(): PopularItemRedditModel
+    suspend fun getMoreData(): Flow<PagingData<Children>>
+}
