+++ app/src/main/java/com/android/lookerpopularapi/util/Const.kt
@@ -1,0 +1,3 @@
+package com.android.lookerpopularapi.util
+
+const val BASE_URL ="https://www.reddit.com/"
