--- app/src/main/java/com/example/lookerpopularapi/data/models/childrenModels/Children.kt
@@ -1,10 +1,0 @@
-package com.example.lookerpopularapi.data.models.childrenModels
-
-
-import com.google.gson.annotations.SerializedName
-
-data class Children(
-    @SerializedName("data")
-    val childrenData: ChildrenData,
-
-    )
