--- app/src/main/java/com/example/lookerpopularapi/data/models/popularItemModels/PopularItemRedditModel.kt
@@ -1,11 +1,0 @@
-package com.example.lookerpopularapi.data.models.popularItemModels
-
-
-import com.google.gson.annotations.SerializedName
-
-data class PopularItemRedditModel(
-    @SerializedName("data")
-    val `data`: Data,
-    @SerializedName("kind")
-    val kind: String
-)
