--- app/src/main/java/com/example/lookerpopularapi/data/source/AppState.kt
@@ -1,7 +1,0 @@
-package com.example.lookerpopularapi.data.source
-
-sealed class AppState{
-    data class Success(val data: Any) : AppState()
-    data class Error(val error_ : Throwable) : AppState()
-    object Loading: AppState()
-}
