--- app/src/main/java/com/example/lookerpopularapi/data/source/remote/ApiSource.kt
@@ -1,8 +1,0 @@
-package com.example.lookerpopularapi.data.source.remote
-
-import com.example.lookerpopularapi.data.models.popularItemModels.PopularItemRedditModel
-
-interface ApiSource {
-    suspend fun getData(): PopularItemRedditModel
-    suspend fun getMoreData(after: String): PopularItemRedditModel
-}
