--- app/src/main/java/com/example/lookerpopularapi/ui/MainActivity.kt
@@ -1,24 +1,0 @@
-package com.example.lookerpopularapi.ui
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import androidx.navigation.NavController
-import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.ui.NavigationUI
-import com.example.redditpopularapi.R
-
-class MainActivity : AppCompatActivity() {
-    private lateinit var navController: NavController
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        val navHost =
-            supportFragmentManager.findFragmentById(R.id.container) as NavHostFragment
-        navController = navHost.navController
-        NavigationUI.setupActionBarWithNavController(this, navController)
-    }
-
-    override fun onSupportNavigateUp(): Boolean {
-        return navController.navigateUp() || super.onSupportNavigateUp()
-    }
-}
