--- app/src/main/java/com/example/lookerpopularapi/ui/screens/main/MainFragment.kt
@@ -1,60 +1,0 @@
-package com.example.lookerpopularapi.ui.screens.main
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.View
-import androidx.lifecycle.lifecycleScope
-import by.kirich1409.viewbindingdelegate.viewBinding
-import com.example.redditpopularapi.R
-import com.example.lookerpopularapi.data.models.popularItemModels.PopularItemRedditModel
-import com.example.lookerpopularapi.data.source.AppState
-import com.example.redditpopularapi.databinding.FragmentMainBinding
-import org.koin.android.ext.android.getKoin
-import org.koin.core.scope.Scope
-
-class MainFragment : Fragment(R.layout.fragment_main) {
-
-    val scope: Scope = getKoin().createScope<MainFragment>()
-    private val binding: FragmentMainBinding by viewBinding()
-    private val viewModel: MainViewModel = scope.get()
-    private var adapter: MainAdapter? = null
-
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        setupMenu()
-        setupOnView()
-
-    }
-
-    private fun setupMenu() {
-        setHasOptionsMenu(true)
-        setMenuVisibility(true)
-    }
-
-    private fun setupOnView() {
-        viewLifecycleOwner
-            .lifecycleScope
-            .launchWhenCreated {
-                viewModel.getMore().collect {
-
-                    adapter = MainAdapter()
-                    binding.rvMain.adapter = adapter
-                    adapter?.submitData(it)
-                    renderData(it as AppState)
-                }
-            }
-    }
-
-   private fun renderData(appState: AppState) {
-        when (appState) {
-            is AppState.Success -> {
-                adapter = MainAdapter()
-                binding.rvMain.adapter = adapter
-                if (appState.data is PopularItemRedditModel) {
-                    adapter = MainAdapter()
-                    binding.rvMain.adapter = adapter
-                }
-            }
-        }
-    }
-}
