--- app/src/main/java/com/example/lookerpopularapi/util/Const.kt
@@ -1,3 +1,0 @@
-package com.example.lookerpopularapi.util
-
-const val BASE_URL ="https://www.reddit.com/"
