+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "RedditPopularApi"
+rootProject.name = "LookerPopularApi"
 include ':app'
